+++ src/test/converter2/issues/gh2929.js
@@ -1,0 +1,15 @@
+/**
+ * @template {number} [T=1]
+ */
+export class NumberManager {
+    /**
+     * @param {T[]} nums
+     */
+    constructor(nums) {}
+}
+
+/**
+ * @template {number} A, B
+ */
+export class EdgeCases {
+}